+++ src/config.lua
@@ -40,6 +40,11 @@ local config = {
   late_biome_damage_ramp = 100,
   late_biome_health_ramp = 100,
   late_biome_start = 5,
+  endless = {
+    health_gain = 0.1,
+    damage_gain = 0.1,
+    gain_type = "exp"
+  }
 }
 
 local configDesc = {
@@ -58,6 +63,11 @@ local configDesc = {
   late_biome_damage_ramp = "Enemy damage ramp for biomes >= late_biome_start",
   late_biome_health_ramp = "Enemy health ramp for biomes >= late_biome_start",
   late_biome_start = "Biome depth to start ramping up the scaling from",
+  endless = {
+    health_gain = "Percentage increase in health of enemies per biome after region 12.",
+    damage_gain = "Percentage increase in damage of enemies per biome after region 12.",
+    gain_type = "Choose between linear or exponential scaling after region 12. Valid values: exp, linear"
+  }
 }
 
 return config, configDesc